requirements.txt
pytorch
numpy
matplotlib
tqdm
requests